Enable processing of TPM 2 Physical Presence requests
This patch enables the processing of a TPM 2 Physical Presence command.
While we do not enable the Physical Presence Interface (PPI) on the OS level,
which would require SMM and ACPI support, we do need it for the processing
of PPI codes that are issued by the UEFI menu and need to be processed
upon reboot.

The first call to Tcg2PhysicalPresenceLibSetupDXE() must be done before
the end of the DXE phase. In this call we create the Physical Presence
and Physical Presence flags variables if they are missing. We also lock
the Physical Presence flags variable, which must be done during DXE.

The subsequent call to Tcg2PhysicalPresenceLibProcessRequest() does
the actual processing of the PPI code and may prompt the user for
confirmation or rejection.

Contributed-under: TianoCore Contribution Agreement 1.1
Signed-off-by: Stefan Berger <stefanb@linux.vnet.ibm.com>