Extract variable locking from Tcg2PhysicalPresenceLibProcessRequest()
We split off the variable cration and locking part found in the current
implementation of Tcg2PhysicalPresenceLibProcessRequest() into
Tcg2PhysicalPresenceSetupDXE(), which must be called in the DXE phase
for the locking to succeed. The existing function may be called once
the console is ready since it may prompt the user for confirmation.

There are no callers of the existing code that would need to be adapted.

Contributed-under: TianoCore Contribution Agreement 1.1
Signed-off-by: Stefan Berger <stefanb@linux.vnet.ibm.com>