Extract variable locking from TcgPhysicalPresenceLibProcessRequest()
We split off the variable creation and locking part found in the current
implementation of TcgPhysicalPresenceLibProcessRequest() into
TcgPhysicalPresenceSetupDXE(), which must be called in the DXE phase
for the locking to succeed. The existing function must be called once
the console is ready since it may prompt the user for confirmation.

Adjust all the callers to the function.

Contributed-under: TianoCore Contribution Agreement 1.1
Signed-off-by: Stefan Berger <stefanb@linux.vnet.ibm.com>